pyproject.toml
[project]
name = "john-pork-refactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "loguru>=0.7.3",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "sqlalchemy>=2.0.46",
]
